Changes in utils.algn_img
